feb_5/requirements.txt
pypdf
sentence-transformers
faiss-cpu
numpy
google-genai